Add files via upload
At the request of Mr. Adil MAALLEM. An initial script that is functional, but not very modular. The new approach to this script allows you to export a specific set of information from a machine. But also to cut the domain name in order to extract a specific part designating the locality of the infrastructure. All with a timestamp and information about the person who carried out it.